pyproject.toml
[project]
name = "documind"
version = "0.1.0"
description = "Production RAG system for private document intelligence — hybrid retrieval, parent-child chunking, streaming answers with citations, and automated RAGAS evaluation CI/CD gate."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "pydantic-settings>=2.13.1",
    "tiktoken>=0.12.0",
    "unstructured[pdf]>=0.21.5",
    "uuid>=1.30",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pytest>=9.0.2",
]
